fix: correct threshold comparison in percentage rate limiter
Prevent incorrect blocking when threshold has fractional values by converting
the integer percentage to float64 instead of truncating float64 threshold to int

- Convert quota.Download.Percentage to float64 before comparison
- Update test to reflect correct behavior for fractional thresholds
- Previous code truncated fractional thresholds to int, causing incorrect blocking
  at specific percentage values

+++ account.go
@@ -1457,6 +1457,6 @@ func CreateDownloadPercentLimitedRateLimiter(client AccountAPI, thresholdPercent
 
 		// Block downloads if usage is at or above the threshold percentage
 		// Return true (allow) only when usage is strictly below threshold
-		return quota.Download.Percentage < int(thresholdPercent), nil
+		return float64(quota.Download.Percentage) < thresholdPercent, nil
 	}
 }
